Convert models.py into package
